[API] vfs: do not use *64() function names
The function names fstat64 lseek64 and ftruncate64 may be defined to map
to other names through the preprocessor in some C library
implementations. This bleeds into the classes and makes debugging harder
and risks name collisions. To avoid this, rename the internal Vfs
functions to their names without the '64' suffix, keeping the arguments
intact.

Change-Id: I1ca6b0cc3cfaa2d11a17ddc5f6f6f4117ab66f7b