libc/ldso/libdl: harmonize redundant defines and sources
Some sources were manually added for each architecture we support to the
architecture specific source list. Harmonize them to be in the global
source list.

Also, the USE_TLS define was used inconsistently between ldso, libdl and
libc. In particular libc did not specify it for riscv, while ldso and
libdl did. Also for sparc / ppc32 it seemed to be specified
inconsistently.

Synchronize the specification to simplify the buildsystem.

Change-Id: I8e8356e33b579642bfc4339a80bda5ffa590e0d8