uclibc/mk: factor out source information
The information which groups of source files belong into which version
of the library is specific to the C library. To separate the build rules
from the information what to build factor these into their own file.

Use a common file for minimal and non-minimal that lives in the concrete
C-library directory. While at it rename UCLIBC_BUILD_MINIMAL into
LIBC_BUILD_MINIMAL to make the build system more libc agnostic.

Change-Id: I71853a988056b72708ffed06e218821a0cc16092