mk: remove MKFLAGS variable
The variable is never set by anyone and also has no implicit automatic
meaning. Recursive invocations of make through $(MAKE) already inherit
the make flags of the invoking make through the MAKEFLAGS special
variable.

Also, the PKGDIR variable is overridden inside the libpthread include
Makefile and thus can be omitted and it can be explicitly set in
Makefile.install for the uclibc-headers since we need to know the
structure already anyways due to the L4DIR set there. So no point in
forcing this from the outside.

Change-Id: I53b72f58411ef87fe977a7f8c6823d81eb86f886