uclibc: Fix truncation problem in malloc
This fixes a bug that can lead to the calculation of a wrong bin `idx`,
which in turn leads to a too small chunk of memory being chosen for the
number of bytes (`nb`) to be allocated. This leads to a fault (or
possibly to memory being written in the wrong location) when using the
offset `nb` on top of the chunk for a write operation.

See ticket #CD-1508

Change-Id: Ief3d48d8f9f9cc003827f820a140e25326546c34