l4re: implement raise()
Let's a thread send a signal to itself. More importantly, it will
provide proper handling of abort() as mandated by the C and POSIX
standards.

Unfortunately, we cannot directly implement raise() in libc_backends
because it is a libc internally hidden function. As a workaround we
forward the call to another function.

See ticket #CD-1507.

Change-Id: I20151ccd93fa4fcc44ff45671b438b0bc4d7945f