Allocation-free libc-minimal, separate libumalloc
This removes the dynamic memory allocation from libc-minimal. In case a
component that uses libc-minimal requires dynamic memory allocation
(e.g. due to the use of C++ exceptions), it can use libumalloc which
implements a basic, self-contained, standard-compliant memory
allocator.

This also changes the dependency on the uclibc_morecore() API to a
custom, simplified libumalloc API.

Change-Id: Icbc68d568ad71ae713ea265e3d73ae742c8b6eda