ldscripts: Fix broken manual make rule generation for crt1.p.o
The manual make rule generation for crt1.p.o is broken, because it
assumes that $(NAME_crt1_reloc) will always refer to an assembler file.
This assumption holds for uclibc, but breaks for musl, which implements
it in C.

Switch to detecting C files by their .c extension and defaulting to AS
otherwise.

Change-Id: I6efb624ce6b73d1c11fbea6c86b7d344d083ae2d