libc_backends: remove ptsname stub
The ptsname() function is not declared in stdlib.h because
__UCLIBC_HAS_PTY__ is undefined. We should not define it either because
that might fool autoconf configure scripts. They test for the presence
of the symbol in libc and do not check for the declaration. Now,
because the symbol was detected, the compilation might fail due to the
missing function declaration.

As we don't have stubs for any of the other pseudo-terminal functions,
we'd better just remove it.

Change-Id: Ie3dd2473dea6afa1a85b324f9a69516027492858