moe: clarify handling of scheduler proxies
Scheduler proxies can be created only on the root factory. All
additionally created factories must not allow the creation of new
scheduler proxies because that would open up a security hole.

The code did this properly but was not really readable in this respect.
Give the member variable the proper name without changing the
observable behaviour.

Change-Id: I551dd98ff1da37bd85fdd934f7884cb2e16c89be