libpthread: Reset pthread_descr pointer in unused UTCB
This is required so that the invalid_handle() and nonexisting_handle()
functions can reliably detect a freed thread handle.

If not reset, the functions will dereference the dangling pthread_descr
pointer still stored in UTCB.user[0], and try to access already freed
memory.

Wasn't noticed with uclibc-ng so far, because it uses malloc/free to
allocate the pthread_descr. The backing memory is freed, but remains in
the malloc memory pool, so the address remains mapped in the address
space and the access triggers no page fault.

Change-Id: I72097168ecd71960984d9eb9bdcef8b4410903dc