cxx: support more than one interface in meta protocol
Walk the tree of base classes instead of just returning the topmost
interface.

Special care must be taken to skip invalid entries in the L4::Type_info
hierarchy. At the leaves of the type tree, there is always an
L4::Kobject that must not be reported. Likewise, interfaces with the
L4::PROTO_EMPTY protocol are not dispatched and should not be visible
either.

Change-Id: Ib4c8687e98ddb1910cd3ef11bedd80399fae5ddb
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>