AVL set: Remove dead code
As far as we can tell, the `struct No_type` forward declaration does
not serve any purpose in cxx::Bits::Base_avl_set::Node, it does not
affect memory layout or any property of the class, it is unused and
it cannot be used by any derived class (since it is private).

It is most likely a leftover that can be safely removed, to avoid any
further confusion.

Change-Id: I65727327be5498a719986089391459d26c930fe4