uclibc-ng: avoid crashes in statical linked binaries when dlopen()
Even in current (1.0.55) releaes of uClibc-ng there will be 100%
reproducible crashes of statically linked binaries (on all kind of
platforms), when calling "dlopen(...)" with wrong or non-existing
.so-files).

This is caused by missing checks on "_dl_loaded_modules" in
"ldso/ldso/dl-elf.c". When "_dl_loaded_modules" is NULL in static
linked binaries, it becomes dereferenced and causes an segfault.

This patch fixes the issue by adding an extra assignment-check for
"_dl_loaded_modules".

Change-Id: I4ee92df1ce0a203bccdfa37e2ec16c11dcb702d0
Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>