l4re/util: fix Vcon_svr read buffer overflow
Per contract vcon_read() can return the input size + 1 to signal more
data being available to read. The maximum allowed input size is
specified as L4_VCON_READ_SIZE, which corresponds to the maximum number
of bytes the UTCB message registers can hold minus the opcode register.

Vcon_svr ensures, the call to vcon_read() gives at most
L4_VCON_READ_SIZE, however, the return value check is insufficient and
may lead to a one-byte buffer overflow.

When calling vcon_read() with L4_VCON_READ_SIZE and there is more data
to read, vcon_read() returns a value of L4_VCON_READ_SIZE + 1. Vcon_svr
takes this value to memcpy data from its local buffer to the UTCB, which
then writes one byte past the UTCB MRs. The value written is whatever is
on the stack behind the local buffer and it overwrites the lower byte of
the field behind the UTCB MRs.
This field holds the base address of this UTCB itself. On x86_64 this
field is used to find this thread's UTCB and is accessible via `gs:0`.

Each call to l4_utcb() / l4_utcb_direct() uses `gs:0` to indirectly
access this field and if the lower byte is overwritten with another
value than zero, the next access to the UTCB will be offset by up to
256 bytes.
Thus, a legal memcpy operation with this offset UTCB address has the
potential to overwrite the whole UTCB-address field behind the UTCB MRs.

In the best case, the service using Vcon_svr crashes. In the worst case,
the UTCB pointer can be manipulated to read/write arbitrary data in the
service's address space.

This commit also fixes a second issue, when vcon_read() returns
L4_VCON_READ_SIZE + 1: the function l4_bytes_to_mwords() returns the
number of UTCB message registers to copy, which for L4_VCON_READ_SIZE
is 62, but for L4_VCON_READ_SIZE + 1 is 63. The message tag computed
in l4_msgtag() also needs to account for one message register with the
opcode, which sums up to 64. However, the message tag specifies
exactly 5 bits for the number of message register, because the message
register array size is 63 fields. A value of 64 masked with 5 bits is
zero.
Thus in the IPC reply of this service, the kernel copies zero bytes
from the services UTCB MRs to the callers UTCB MRs.
When the caller now checks its UTCB MRs, it finds its own input values
and wrongfully interprets them as output values. Which in turn can lead
to a crash of the caller application.

Fixes ticket #CD-1649.

Change-Id: I0965a70f625c85aacc2441333bd13bc9379ae106
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>