rm: add notion of private and anonymous regions
Anonymous regions are not backed by a dataspace. Instead, the region
manager provides anonymous memory itself in the requested region.
Because there is no capability to map the content anywhere else, always
free on detach.

Private regions are copy-on-write mappings. Writes to the region are
not written back to the underlying dataspace. This can be combined with
anonymous mappings to provide proper fork() semantics.

Change-Id: Ic2c7f9ce79df75637aeb268e3a261841b08abd75
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>