[API] Implement enum bitwise operators as universal templates
Instead of defining bitwise operators on enumeration types using
a macro, provide them as universal templates.

For an enum type T to opt-in for these operators, there needs to be
a declaration of a function

  void enum_bitops_enable(T)

available for the argument-dependent lookup (the function is never
actually called, thus no definition is required and no code is
generated).

In addition to the previous implementation, the compound assigment
operators '&=' and '|=' are also provided.

New set-difference '-' and '-=' operators (a shorthand for intersection
with a negated second argument) can be used to clear specific bits from
the enum value. There is also the to_underlying() helper function.

Signed-off-by: Martin Decky <martin.decky@kernkonzept.com>
Change-Id: I6bcd79e4f1bc3fab6daa8aa7f24dad4bf111120a