l4re_itas: make nofpu
We're about to handle most page faults in the context of the thread.
We'd better not disturb the FPU state which is not saved by the kernel
when jumping to the page fault handler.

Note that the signal handler trampoline code runs in the context of the
application threads and therefore needs to save/restore FPU registers.
Add a note to the platform-dependent signal handler trampolines.

Change-Id: I0b4f35ffdf7b3588d4119af007c8a691fc7c92f0
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>