musl: Workaround limitation of L4Re mmap implementation
According to the POSIX specification of mmap it shall be possible to map
a region of memory that is larger than the backing file. Musl makes use
of this feature when loading shared libraries, but unfortunately the
current L4Re mmap implementation does not support it.

As a workaround, we use the alternative code paths that musl implements
for DL_NOMMU_SUPPORT.

Change-Id: I719b7242fcebcb8d7493ab7d5f6c12671e935f9b
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>