libc/buildsystem: Add feature for generated headers
Add a new class of headers that are not symlinked from the source dir
but instead need to be generated. All headers listed under the GEN_
category must have an associated Make rule defined in the
target_headers.lst (or a file included from there) that describes how
they can be built.

Change-Id: I153e442fa94addf7d9190c0dc34f19e0bde3afe0
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>