Private and anonymous regions in moe and ITAS
While both components provide the same semantics in the end, their
implementation differs substantially.

In moe, being the principal provider of anonymous memory in the system,
the implementation is quite simple. For private mappings of dataspaces,
a copy_in is done automatically that handles already all the details.

The ITAS implementation is a bit more involved. For anonymous mappings,
a largs, shared dataspace is allocated. This mirrors the current
implementation in l4re_vfs. For private mappings, a moe dataspace of
matching size is allocated together with a bitmap that tracks which
pages were already copied from the underlying datspace to the anonymous
shadow copy. Whenever a write page fault happens in a private mapping
region, the page is copied into the anonymous dataspace. From there on
the page from anonymous datspace is used.

Change-Id: I29dbdeefc8e311d6caac25d2f9f9d2c38e7aa430
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>