libc_backends/l4re_vfs: Guard uclibc specific flags
The libc-symbols header and the NOT_IN_libc preprocessor define are
uclibc-ng specific and configure the headers to make certain defines
available and ensure the presence of some symbols (in particular the
xxx64 variants of file functions are important to us).

Guard these flags to only pull them in when our library is uclibc-ng.

Change-Id: I8111c39b7da9dbc4c03cb7a4d9006f5c89a1db44
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>