libc_backends: derive noexcept specifiers
Wether a function declared in a C header is noexcept depends on the
library that provides the C header and, further on the compiler, as
different compilers treat C functions that are not declared nothrow
differently.

Thus, in order to not have to rely on one or the other, derive the
noexcept attribute for all functions that implement C library
functionality.

Change-Id: I44307031131ca12232e7d648fca30ead661a70d3
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>