libc-headers: Always update headers symlinks if the new target differs
When we add a new override header, then there is nothing telling the
buildsystem to update the symlinks. To force this we need a tracker that
gets only updated when one of the source headers is newer than the
tracker. We use $? for that which contains the vpath resolved
prerequisites which have a newer timestamp than the target.

Change-Id: I4305e4d27f0edb8c8bff8f6ce84520a5b4bd6200
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>