musl: Introduce execute in pthread manager request
The REQ_EXEC_IN_MANAGER request replaces the more specific
REQ_FOR_EACH_THREAD request, with the latter now being implemented using
the former.

The REQ_EXEC_IN_MANAGER request allows to execute more complex
operations in the context of the manager, which is required to implement
dlopen for musl.

Change-Id: I3759ade617b944069b374f657a65fba10374c104
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>