l4re: fix possible nullptr deref in L4::Cap_base
It is possible to construct an L4Re::Shared_cap with a nullptr for the
Cap_alloc interface. This is handled correctly in the free() method.
Adapt the copy() method too.

The _ca member will already contain the right capability allocator
pointer. The L4::Detail::Shared_cap_impl class calls the assignment
operator before calling copy().

Change-Id: I78bca2aa73dcc0f6de31c28a1abfc32ae69fad3d
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>