libc: set global KIP pointer when linked shared
If the application is linked shared, the l4_global_kip pointer is
coming from the ld.so. Initialize it with the value from the AUX
vectors. Otherwise it will have the default value and, if that is
different from the loader chosen one, will cause a crash on the first
KIP access.

Change-Id: I68689edf9ff0b634baedc9e28bf526a8440560b4
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>