l4sys: l4_timeout_from_us() returns a value greater than input
Up to now, l4_timeout_from_us() returned a 16-bit timeout value that
was the next lower value expressable with 5-bit exponent and 10-bit
mantiassa.
Since a timeout is generally defined as a minimum waiting period, we
change the implementation to return the next higher value than the input
value.

This leads to a change at the upper limit of the value range: 2^41 - 1.
While the old code could handle this input, the new code exceeds this
range and overflows the 10-bit mantissa.
So we additionally add bounds checks to the code handling the upper
limit and note it in the interface documentation.

Change-Id: I8ea4ff22480707e01f734f6035cb89a6225add92
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>