libpthread/{signal,mutex}: synchronize musl/uclibc
The bits/sigthread.h header is included from signal.h in uclibc in all
cases we care about. It is also not there for musl. So remove the
include from the signal.c impl. in uclibc-ng as we did for musl.

The sigprocmask definition is also not an issue. Keep it to be in sync
with musl.

While at it synchronize some of the indentation to reduce the diff.

Change-Id: I09e023bce455aff95638e920845775ac19d75ddf
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>